+++ core/src/test/resources/com/google/errorprone/bugpatterns/BadShiftAmountPositiveCases.java
@@ -23,16 +23,15 @@ public class BadShiftAmountPositiveCases {
 
     public long testEquality(int x) {
 
-        int result = 0;
+        long result = 0;
          
-        //BUG: Suggestion includes "x"
+        //BUG: Suggestion includes "(long)x >> 32"
         result += x >> 32;
-        //BUG: Suggestion includes "x"
+        //BUG: Suggestion includes "(long)x << 32"
         result += x << 32;
-        //BUG: Suggestion includes "x"
+        //BUG: Suggestion includes "(long)x >>> 32"
         result += x >>> 32;
         
-         
           return result;
       }
